requirements.txt
flask
pdfplumber
requests
readability-lxml
beautifulsoup4
fpdf
ebooklib
gunicorn
